# docs/api.yaml
openapi: 3.1.0
info:
  title: WeFixIt HTTP API
  version: 1.0.0
  description: Session-authenticated endpoints used by the server-rendered application.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wefixit_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
paths:
...
  /booking/book:
    post:
      summary: Create a service request
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '201': { description: Booking created }
        '400': { description: Validation failure }
        '403': { description: Customer role required }
  /booking/cancel-booking:
    post:
      summary: Cancel an eligible owned booking
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '200': { description: Booking cancelled }
        '409': { description: Invalid lifecycle transition }
...
  /jobs/{bookingId}:
    get:
      summary: View an authorized job record, notes, photos, and completion confirmation
      security: [{ sessionCookie: [] }]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Job record rendered }
        '404': { description: Job is not visible to this user }
  /jobs/{bookingId}/pdf:
    get:
      summary: Download the authorized job card as a PDF
      security: [{ sessionCookie: [] }]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Job-card PDF
          content:
            application/pdf: {}
...